# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Maximizing Your Chat with AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /flows:
    get:
      operationId: listFlows
      tags: [flows]
      summary: List chatbot flows for the current owner
      responses:
        "200":
          description: Flow list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlowSummary"
    post:
      operationId: createFlow
      tags: [flows]
      summary: Create a new (empty) flow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlowCreateInput"
      responses:
        "201":
          description: Created flow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flow"
  /flows/import:
    post:
      operationId: importFlow
      tags: [flows]
      summary: Import a flow from a backup export (creates a new inactive flow)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlowImportInput"
      responses:
        "201":
          description: Imported flow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flow"
  /flows/{id}:
    get:
      operationId: getFlow
      tags: [flows]
      summary: Get a flow with its full graph
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Flow detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flow"
        "404": { description: Not found }
    patch:
      operationId: updateFlow
      tags: [flows]
      summary: Update flow name and/or graph
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlowUpdateInput"
      responses:
        "200":
          description: Updated flow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flow"
        "404": { description: Not found }
    delete:
      operationId: deleteFlow
      tags: [flows]
      summary: Delete a flow
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }
        "404": { description: Not found }
  /flows/{id}/activate:
    post:
      operationId: activateFlow
      tags: [flows]
      summary: Set this flow as the single active flow for the owner
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Activated flow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flow"
        "404": { description: Not found }
  /flows/{id}/deactivate:
    post:
      operationId: deactivateFlow
      tags: [flows]
      summary: Deactivate a specific flow
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deactivated }
        "404": { description: Not found }
...
components:
  schemas:
...
    FlowNode:
      type: object
      required: [id, type, position, data]
      properties:
        id: { type: string }
        type:
          type: string
          enum: [trigger, message, question, end, ai, products]
        position:
          type: object
          required: [x, y]
          properties:
            x: { type: number }
            y: { type: number }
        data:
          type: object
          properties:
            matchType:
              type: string
              enum: [default, keyword]
            keywords:
              type: array
              items: { type: string }
            text: { type: string }
            imageUrl:
              type: ["string", "null"]
              description: "Message/Question only: optional image URL (http(s) or /api/media/...) sent with the text as caption."
            options:
              type: array
              items:
                type: object
                required: [id, label]
                properties:
                  id: { type: string }
                  label: { type: string }
            strictOptions:
              type: boolean
              description: "Question only: re-ask question on off-option replies instead of handing off to AI."
            strictRetryMessage:
              type: string
              description: "Question only: when strictOptions=true, this message is sent before the question is re-asked. Leave empty to skip."
            aiRephrase:
              type: boolean
              description: "Question only: when true, the question text is rephrased by AI (same meaning, varied natural wording) each time it is sent, so it doesn't feel like a canned bot message. Answer options are never rephrased."
            productIds:
              type: array
              items: { type: integer }
              description: "Products only: list of product IDs to send (image + Nama/Kode/Harga caption)."
            aiInstruction:
              type: string
              description: "AI node only: extra instruction appended to the global AI Studio system prompt while this node's AI handoff is active (during the flow cooldown). Empty = use the global prompt only."
            knowledgeIds:
              type: array
              items: { type: integer }
              description: "AI node only: restrict the AI's knowledge-base reference to these specific knowledge entry IDs while this node's handoff is active. Empty = use all of the owner's knowledge base."
    FlowEdge:
      type: object
      required: [id, source, target]
      properties:
        id: { type: string }
        source: { type: string }
        target: { type: string }
        sourceHandle:
          type: ["string", "null"]
    FlowGraph:
      type: object
      required: [nodes, edges]
      properties:
        nodes:
          type: array
          items: { $ref: "#/components/schemas/FlowNode" }
        edges:
          type: array
          items: { $ref: "#/components/schemas/FlowEdge" }
...
    FlowSummary:
      type: object
      required: [id, name, isActive, channelIds, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        isActive: { type: boolean }
        channelIds:
          type: array
          items:
            type: integer
          description: "Channels this flow is assigned to. Empty array = global (all of the owner's channels)."
        updatedAt: { type: string, format: date-time }
    Flow:
      type: object
      required: [id, name, isActive, channelIds, graph, createdAt, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        isActive: { type: boolean }
        channelIds:
          type: array
          items:
            type: integer
          description: "Channels this flow is assigned to. Empty array = global (all of the owner's channels)."
        graph: { $ref: "#/components/schemas/FlowGraph" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    FlowCreateInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        channelIds:
          type: array
          items:
            type: integer
          description: "Channels to assign the flow to. Omit or empty array = global (all of the owner's channels)."
    FlowUpdateInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        graph:
          $ref: "#/components/schemas/FlowGraph"
        channelIds:
          type: array
          items:
            type: integer
          description: "Channels to assign the flow to. Empty array = global. Omit to leave assignments unchanged."
    FlowImportInput:
      type: object
      required: [name, graph]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        graph:
          $ref: "#/components/schemas/FlowGraph"
...